Move some shared methods in to root of tree mod
